+++ app/src/main/java/hu/bme/aut/android/srm/model/FermentTemp.kt
@@ -3,6 +3,6 @@ package hu.bme.aut.android.srm.model
 import java.io.Serializable
 
 data class FermentTemp(
-    override val value : Int? = null,
+    override val value : Double? = null,
     override val unit : String? = null
 ): Serializable, Temp
