+++ app/src/main/java/hu/bme/aut/android/srm/model/Temp.kt
@@ -1,6 +1,6 @@
 package hu.bme.aut.android.srm.model
 
 interface Temp {
-    val value : Int?
+    val value : Double?
     val unit : String?
 }
