+++ app/src/main/java/hu/bme/aut/android/srm/model/TempStep.kt
@@ -3,7 +3,7 @@ package hu.bme.aut.android.srm.model
 import java.io.Serializable
 
 data class TempStep(
-    override val value : Int? = null,
+    override val value : Double? = null,
     override val unit : String? = null,
-    val duration : Int? = null
+    val duration : Double? = null
 ): Serializable, Temp
