+++ app/src/main/java/hu/bme/aut/android/srm/model/WaterVolume.kt
@@ -3,6 +3,6 @@ package hu.bme.aut.android.srm.model
 import java.io.Serializable
 
 data class WaterVolume(
-    val value : Int? = null,
+    val value : Double? = null,
     val unit : String? = null
 ): Serializable
